A more elegant weapon from a more civlised age
adds what is effectively a rebalanced sword.

instead of being in sword family it's in two handed, and allows some tricks.

Like most two handed weapons it gains a damage boost if wield. At 0.3 sharpness it will be as powerful as machete of similar material. Minus the cleave effect. If held with something in your other hand it does 0.1 sharpness, a third of the machete.

Also the parry ability ONLY works if wielded. So you can't use it with a shield or a second parry item. You can't even parry if you have a gun or pen in the other hand.

so pros
can parry

cons
needs an empty hand to parry
if used one handed is 1/3 as powerful as stock gear
starts out as steel and needs mats to upgrade