Assistant alt title for Intern
Assistant was renamed as a job to attempt to get away from the general SS13isms that come with the default position.

I think though that, as a title, it fits very well with Intern, especially considering it isn't the default role, or the default title for that role. 

Basically, it's what I'd use for playing intern, since when I'm playing intern I'm generally hopping on to be a helper who doesn't necessarily have responsibilities unless I volunteer for them.

And if that doesn't sound like 'someone who assists', I don't know what does. SO! (I thought about doing '[department] assistant' for all the different departments except command (since Command Assistant already exists as an alt title) but I sort of thought better of it.