Allowing micros to crawl into storage items
Fixed a bug where a micro player could pick themselves up out of storage items if they used alt+click to open it. This would free them, but create an invisible, empty holder item in their hands which could then be placed in storages.

Added the ability for micros (size 25%) and observer mice (size 100% or lower) to drag their sprite over storage items in order to climb inside. Same as if a player were to pick them up and store them.