Auto-Resleever accounts for brains
If you click on the auto-resleever while you still have an active synth brain somewhere, it will ask you if you want to delete that brain. 

This is to prevent harvesting brain parts from dead people or other weirdness for brains getting turned in after they are killed and auto-resleeved.