Xenochimera Feral Reduction in Darkness OR Belly
Xenochimera will now reduce ferality at a tickrate of 1 per Life tick while inside a belly or in darkness, same as the hunger reduction if they're full.
They will not experience hallucinations and will get safety messages while inside a belly, as well as the status indicator confirming they're safe even in a lit environment.

In total:
If sated, feral will reduce by 1 per Life() tick.
If in darkness _or_ a belly, feral will reduce by 1 per Life() tick.
A feral of 220 will take 110 seconds to clear if satiated, not in a belly, and in light.
Alternatively, it will take 55 seconds to clear if satiated **and** either in a belly _or_ in the darkness.

Upstream port of https://github.com/CHOMPStation2/CHOMPStation2/pull/3823
and
https://github.com/CHOMPStation2/CHOMPStation2/pull/3906