Adds Map Templates Folder
This folder exists to allow Event Management staff to publicly post event files, or for users to contribute them for use by EMs later, as well as a place to point new contributors towards - for instance, the Complete Room Checklist.dmm has everything a maintainer will check is needed for a room.
Templates is for any sort of template an EM or other staff might use on the fly that is not necessarily related to events.

Refer to (and update) the readme.md if any of this changes.