Adjust default prefs
Makes it so the stumble/slip/drop vore prefs are on by default, since you have to have the spontaneous vore option enabled for them to work to begin with. 

Also makes both spontaneous vore options off by default for consistency. (And moves the enabling of the pred option to the simple mob file since that's what that was for anyway)