Mob Examine
Introduces the 'Mob Examine' verb, which lives in the right click menu and the IC tab. 

This verb is functionally identical to regular examine, except:
-Mob Examine only works on mobs
-Mob Examine will allow you to examine people your character cannot technically 'see', such as in cases where you are in a belly, or in a micro holder, or you are in a locker.

They do need to be visible to the turf whatever you are in occupies though, so it's not like you'll be able to xray people through walls, or inside of other people. However, this should enable people who are being held in hands or in tummies to examine the people around whoever is holding them for purposes of checking prefs and descriptions.

As a note, I did try to add this on to the basic examine verb to make it more seamless, but how examine gathers its targets is not something I really understand, and the way it works totally fails out before it even gets to the code in the verb, so I can't just tell it to run this proc if they're in a belly or a holder, and I didn't want to break it by changing it. SO! Separate verb.