TGUI Bundle Update Label
It could be very useful for reviewers and developers to already see in the overview if one or multiple PRs are updating the TGUI bundle file at the same time (which could cause merge conflicts).

The interface files themselves are easier to be merged compared to this compressed compiled outcome.