Silicon sensors now use omnihud
Changes the toggle_sensor_mode proc to toggle omnihud panes on or off. It gives the same panes as the AR-Bs do, so you get all the relevant information you could want.

Mostly messes around with silicon files but also changes the examine behavior for records on borgs to conform to the new system.