New pAI verb
Screen Message allows a pAI to display a message on their screen.

This makes it so the message can be read if you open the card interface, and, if you happen to be holding the card in one of your inventory slots that isn't a container of something else (or if you have them in your PDA or tummy), then it will print the message in the chat.

I decided to let it work for the tummy thing despite not making logical sense because people use bellies for all kinds of stuff, and it's not like you can powergame this so WHATEVER!!!!!!!!!!!!!!!

Logs anything you write to this into say, annnnd. Yeah!

This should let people who like to play mute pocket friends (me) quietly talk to the person holding the card, even when comms are down! (wow!)