Mouse Ray
It turns things into mice (if they have the spontaneous TF preference enabled)

Also wiggles a few things around to allow for some shenanigans with this, namely:

if you are TFed into a thing, damage you take is relayed to your original body proportionally to the comparable HP. 

MEANING, if you have been TFed into a mouse, you have 20 HP. Your human body has 100. You have 1/5th the HP as a mouse. If you as a mouse take 5 damage, your human body takes 25. Be careful!

Additionally, if you as a mouse, or your human body dies, both die. 

Your TFed body being destroyed undoes the transformation. (But also you're probably dead at that point anyway) This allows for, in the case of digestion, your equipment to go where it ought to go.

Lastly, the building blocks are in place for this to be fluffed into other kinds of metamorphosis rays, should staff feel like doing such. I will probably also trick out the /admin variant of this with a bunch of mob types tomorrow when I'm not sleeby