Makes vore replicator functional. Fixes bugs.
- Makes the vore replicator functional (People can be turned into simple mobs)
- Adds a sanity check to make sure the person inserted into the replicators isn't already transformed.
- Makes it so you can turn player-spawned mice into clothing and it'll treat them like a micro (Still respects prefs and asks for the mouse's confirmation.)
- Fixes a bug where if someone was eaten by someone that was transformed into clothes they would be sent into the void.

Micro Transforming Replicator

Creates a heavily modified version of the xenoarcheology replicator.
One version spawns simple_mobs and the other spawns items.
Creates a proc that allows possession of objects.
Makes both versions able to be found by xenoarcheology.

Features:
Alerts people when they pick up an item that is is a possessed item. Impossible to ignore popup.

Makes it so people TF'd into items can revert back to their normal, pre-TF form.

Makes OOC escape functional (An in-game obtainable escape is still needed to be coded in. That's next on the todo list.)

Gives FAIR WARNING to the person being shoved into the replicator that, hey, you're going to be turned into clothing/an item and possibly used as such. Gives them a second warning as well.

Makes the item's eatable&digestable variables adjust depending on the person's variables before TF'd.

Makes it so the 'voice mob' is properly deleted when digested.

Fixed a bug with the replicator allowing you to put grabs in it. (This will be edited to allow for stuffing people proper in it.)

Makes the medical mouse ray undo TFs.

Additionally, makes it so dead people can't be shoved into the replicator.

TGUI and makes the clothing machine function better.

- Removed a debug log I left in the mouseray while testing.
- Adds a note to the clothing replicator's description that indicates it's an object/clothing replicator.
- Allows you to grab and shove people into the replicator without them being a micro. (Still get the same warning checks as a micro)
- Makes the replicator spit out/drop all of the person's items when someone is shoved into it.
-Fixes a TGUI bug with the clothing replicator that caused it to clash with the normal replicator.