Make build mode not crash your client
If you right click on advanced build mode and then push cancel, or just push confirm without entering anything, it tries to build a list of every possible thing in the game, and usually crashes your client trying to display that list.

This makes it just not do anything if you don't give it an input. If you want to have a list of everything in the game you can probably still do that if you just enter / or whatever.