Makes ore boxes, bags, and the processing machine use lists.
- Makes Ore boxes, ore bags, and the processing machine all use lists.

Previously, ore boxes would constantly add ore to the contents of the ore box.
This meant that if a miner was constantly adding ore to a crate that has thousands of ores in its contents list, the contents list would be updated numerous times in succession.

Updating a list with 10000 ore 25 times in a single tick makes the server cry.

Making it use lists completely eliminates this problem and makes the entire process of ore more efficient. It simply adds the amount of ores to a list and transfers them between satchel/bag/processing machine.

Using lists means the infinite mining satchel can actually be infinite without worries of constantly dropping & picking up ore and lagging the server.

This has been tested and works.

The only problem however is that I need to figure out a way to make mining bags able to know if they're full or not. Currently they can hold an infinite value of ores.