Ghost default swap
Admins are going to more than likely want to aghost if they slam the ghost button and then the enter key.
Players trying to ghost might not know better, so let's default to 'Stay in body' so they don't slam enter and make a mistake. Make it so they have to consciously click 'Ghost'

🆑
- tweak: Made it so the ghost command defaults are less disruptive.
/🆑