Allows you to select if you want to heal brute or burn via surgery
Instead of shoehorning the player into healing the most damaged type.

This mechanic was also poorly explained and I have people DMing me about "This surgery is bugged it won't let me heal XYZ" when it was actually just a old-code blocking you from healing the lesser damage. 

With the new surgery, this is a thing of the past and can be safely removed.