Adds Anesthetic Tanks to Medical Gripper
A small adjustment that allows medical borgs to take and place anesthetic tanks ONLY. Specifically no oxygen/nitrogen/phoron etc.

This isn't a power spike, but in the case of surgery/trauma borgs allows them to do their job still in the rare event a wall-mounted anesthetic pump is unusable.