Adds ability to customize discomfort
Both pre-made and custom species may now choose up to 10 messages each for warnings while overheated or cold.

If there are no custom messages, game falls back to defaults (which, if the pre-baked species lacks fluff, uses human)

There is a way to reset to defaults.

It is possible to go back and edit any of the 10 messages. The editor shows you the previously set message with an option to cancel out (this avoids over-writing the previous message).