Even more speech bubbles
This update adds even more speech bubbles. These include a 4 animal themed ones: latte fox, blue fox, red panda and blue panda. It also includes 9 occupation themed ones. I know that some people consider speech bubbles to be accents (not that there is anything prescribing them to this), but these can also be used to represent speaking in certain capacities, such as with the authority of a doctor or officer, or performing an investigation. Mostly though, they're just for fun.