adds: New POI loot: broken drone brain!
This is implemented to act as drops by combat drone enemies in the "phoron tanker" POI. I am commiting separately as this is the generalized form to be re-used and distinct enough from mapping.

This variant should not be seen by players. Either view-variable edit examine_canalyzer/multitool/canalyzer_printed or edit in mapmaker or define it in code thru a subtype

If making it spawn from a mob, you'll need to make a subtype. When defining a subtype, New() can be used for dynamic naming like use of drone_name in examine stuff.