adds: ability to make automated staff request in fax TGUI
Adds a new jsx element to the login screen when using fax machines, a button in the top right corner which when pushed calls the request_roles() verb, and thus initiates the auto fax dialogue for the player.

This element has a tooltip to explain what it does when hovered.

Furthermore, for when EITHER the fax machine has an ID && logged in, BUT no paper is inserted OR paper is inserted && target destination is part of admin_departments (centcom, job board, supply) - it displays a blurb about automated staff requests with some semi-IC, semi-OOC text and the same button.

adds a new general functionality to LoginScreen.js

Adds a new JSX element to the LoginScreen file that takes machine type as its prop and returns a per-machine relevant element. This is needed for the fax functionality, and already has it defined.