Offmap housecleaning
Tweaks a couple things relating to gateway and redgate maps

Renames all of the map templates to have Gateway or Redgate prefixes, that way it's more clear what they are intended for when referred to in game

Removes 'Abandoned Island' from the redgate rotation pending a redesign. Also touches up a couple things with it.

Makes a 140x140 compatible version of the zoo gateway map. Does not add this to rotation for now, I would like to further revamp this later.