Fancy!!!
Makes a few changes to things, most of which are functionally identical to before.

Changes the shadekin empathy color to something unique, and more visible in dark mode than the changeling color. Also gives it a special font.

Makes a slightly lighter color as an alt color for dark mode

Makes it so that hivemind languages don't HAVE to be italics, but makes it so all the existing hivemind languages minus shadekin empathy are italicized. (The font I picked for empathy was hard to read while italics)