Map tweaks!
Moves the redgate on the tether and spices up the area around it!

Also removes all manually set d tags for wires across all of the main station maps, the talon, and the fuel depot. They PROBABLY still exist in some shuttles and stuff but I'm not going through literally every map file for it. >:O

... or am I...