Updated Docker file
Changes:

- Replaced the dependency on the tgstation Byond build by integrating their docker file from https://github.com/tgstation/byond-docker/blob/master/Dockerfile directly to no longer being reliant on prebuild Byond images. Any major and minor version of Byond can now be selected directly.

- Changed the remote origin of the rust-g repository to use the vorestation fork instead of the original tgstation repository.

- Changed the rust_g output of the compiled file to use the new format and direct location which the game looks for first.

- Swapped the copy order of the vorestation files and the librust_g.so.

The latter change was necessary as during the copy command, the vorestation pre-compiled librust_g.so gets copied over. This one causes issues though. (Seemingly compatibility as neither Log_Write nor File_Write works). As the game first looked in that directory, the rust_g file wasn't used and to use the rust_g file, the librust_g.so has to be deleted. It's much easier to simply overwrite the copied over librust_g.so with the newly compiled one.