Updates to fantasy redgate
Added a room with a bigger vampire and good fantasy loot behind a locked door that can only be opened with a newly hidden key. Alongside a few turfs, mob and spawner to support this.

Fixed https://github.com/VOREStation/VOREStation/issues/15416 by adding sprites to open dungeon walls, other sprite fix was in a previous PR.

Added a note at the entrance of the redgate asking people to not bring weapons home.