New Mob: Scel
Added a new mob, Scel, created by Kira72.
Comes in 5 different colours. This mob is relatively rare in random mob spawners. It's a retaliate mob with a detailed belly. It looks like a snake-like creature with black flesh and neon liquid visible beneath patches of it. When attacking, it has a 50/50 chance of using a special attack of either a tongue grab or lunge. Added to the maint pred options too.

Fixed a missing item in Rain City.

Added a new admin spawn map which is literally just the Rascal's Pass shuttle in space.

Removed the great wolf catalogue entry from the crypt drake.