tweak(paper): Changes bundle naming logic to be general
Used to be bundles would get renamed based on hardcoded default values of "paper" or "photo". Using the Initial() procedure, we now check for defaults at compile time to be able to adapt to changes in paper/photo names.

This was done to enable checking if a bundle is using its default name or not.