add(fax):  check to prompt renaming when sending to admin
Adds a check on sending fax to admin departments that prompts the user to either rename, cancel or continue as is when they try to send a fax using the initial() (compile time) value. Continuing works as it did, cancelling ceases the operation and renaming renames it.

Also adds a check for when using bundles to check if the bundle title matches either page 1 or page 2. We only check page 1/2 due to bundle naming logic on creation.

add(fax): Adds a buton to interface to rename fax message

Renames the bundle/paper. Also includes a tooltip to explain what it does and why you should do it.