Size Toy Fun
This PR adds three new functions to existing size toys:

Added a new variant of the size collar that is created by using a wrench to break the crystal. This prevents the user from changing the size manually, and triggering the collar with a signal flips the wearer back and forth between a random size and their original size.

Added a new verb "Spin Size Dial" to size guns to spin the dial to a random size. Found in the object tab or on right click.

Added an alt-click effect to gradual size guns that toggles a size-trading mode. With this active, when you shrink a target you grow at the same speed, and vice versa.