Closet vore and Tents (#15808)
* Closet vore and Tents

Added a new verb to the ability tab when interacting with closets called "Vore Occupants". This allows you to eat another person when both of you are inside of the same closet.

Added tents as a subtype of closets.

* Crafting

Added the ability to craft tents with 10 cloth.

* Ghost check and change category

Added a check to make sure user is /mob/living.

Changed category from Abilities to Object because having no abilities means that your tab changes every time you move past a locker.