Jungle Redgate and New Vore Mobs (#15889)
* [WIP] New Vore Mobs

Added new vore mob: Cave Stalker, a 6 legged canine/feline like creature that charges at prey from a distance to pounce them. It immediately stops on the spot if spotted and then retreats quickly if stared down.

* Fixed missing file

* More mobs, still WIP

* More stuff! [Still WIP]

Added another mob called the sonadile that works based on sound similar to the lurker.

Added a bunch of mobs to the maint pred list, including the kururak because that's a cool one.

* Update vorestation.dme

* Update sonadile.dm

* Another mob and stalker tweaks, still WIP, need to add flavour text for all mobs

* Update devil.dm

* Update stalker.dm

* Even more maint pred options

* More Capture crystal mobs

* Finishing touches, no longer WIP!

* Update vore_raptor.dmi

* Added the jungle map redgate

Also resolves a runtime with the scrubble.

Forces jungle redgate only for the first few days at least.

* Update jungle.dmm

* Update jungle_underground.dmm

* Update paint.dm

Fixes paint runtime