Bluespace Crackers (#16062)
* Bluespace Crackers

Added a bluespace cracker item. It works similarly to a normal cracker, in that one person offers it to another to pull the cracker together. When pulled, the winner is rewarded with a toy, a joke, a cheap paper hat and a powerful bluespace effect!

The effects include:
Shrinking to 25% scale
Growing to 200% scale
Having a drug like effect colour their screen.
Making them invisible for one minute.
Knocking them over.
Teleporting them into the losers active belly (if prefs allow!)
Or giving them a bunch of money.

They can be found in the bluespace random item pool and rarely as an item in maint!

* Colourful hats

Added random colours to the paper crown hats

* Added the crackers and backfiring sizegun to trash piles

* Can be rigged by VV now