Butterfly nets (#16088)
Added butterfly nets. These nets can be wafted across a tile, much like a fishing net, to catch butterflies (and people of scale 50% and below)! Using them on themselves either drops the contents, or if the inactive hand is empty, pulls out the contents into the free hand.
They can be crafted from cloth, found in the botanists locker, or rarely found in maintenance.