Emergency door release (#16559)
Added an extra button to all station dorms, an emergency door release. This can be pulled by the occupants to immediately unbolt and open the door even if it is unpowered.

Why we need this:
During deadshifts, it's somewhat common for players to go and scene in a dorm and lock the door. Then if the power goes out, they generally need to ahelp to get out, because they'll have bolted themselves in. Obviously there are a few ways to solve this, such as special doors, making the room always powered or making a variation of the lock itself that doesn't use power.
However, I felt that having an extra lever allows people to keep whatever vibe they want as for power in their scenes, has a cool vibe itself, and I also have plans for how I can use this specifically in events.