Adds Catborg, Kittyborgs, and Gryphborg (#16910)
* Adds Catborg, Kittyborgs, and Gryphborg

Adds Catborgs, Kitty borgs, and Gryphon Borgs

* Makes the sprite the correct path

Clarifies them as large/small instead of catborg again.

Also moves them to /robot

* Moves drakeborg to /robot/resources

It isn't actually used anywhere, but it's a good idea to keep all the robot sprites together.

* Changes things to sleeper. Gives sec/med sleeper colors

* Fixes tummy disappearing

* case

* Gets rid of charname for whitelist

Because if you want to have a borg with multiple names, you'd have to have them added numerous times.

If someone is ALREADY whitelisted for a borg sprite, why not allow them to have multiple names for that sprite?

* Fixes ckey, adds attribution

* Update ATTRIBUTIONS.md

* Update ATTRIBUTIONS.md

Changes license of kittyborg_spotteh_custom.dmi at request of license holder.

* Update ATTRIBUTIONS.md

* Removes ND sprite

This will be a separate PR due to licensing

* undef