Byond 516 warning message (#16988)
Added a warning message if clients are older than 516, warning them that the game will soon update to version 516 and that they will experience issues in the meantime, and eventually not be able to connect at all.

This really should be merged alongside the TGUI core update, as that will cause UI issues for clients still on 515.